requirements.txt
flask
flask_restful
numpy
pillow